chore: Moved auth into the core folder, and fix any import issues
