+++ src/frontend/src/app/core/auth/components/sign-up/sign-up.css
