+++ src/frontend/src/app/core/auth/models/user.model.ts
